Share documentation navigation schema types
